requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
jinja2>=3.1.4
python-multipart>=0.0.9
pillow>=10.3.0
httpx>=0.27.0
python-dotenv>=1.0.1
instaloader>=4.15
